Adds authorized profile operation

+++ backend/src/controllers/profile.js
@@ -5,8 +5,8 @@ const createProfile = async (req, res) => {
 
   try {
     const data = { owner: username, ...req.body };
-    const { _id } = await profileService.createProfile(data);
-    return res.status(200).send({ id: _id });
+    const { _id, name, numbers, whatsapps } = await profileService.createProfile(data);
+    return res.status(200).send({ id: _id, name, numbers, whatsapps });
   } catch (e) {
     console.log(e);
     return res.status(422).send();
@@ -19,31 +19,41 @@ const getProfile = async (req, res) => {
   try {
     const profile = await profileService.findProfile(id);
 
-    if (profile) {
-      const { _id: id, name, numbers, whatsapps } = profile;
-      const data = { id, name, numbers, whatsapps };
-      return res.status(200).send(data);
-    } else {
+    if (!profile) {
       return res.status(404).send({ error: 'profile does not exists' });
     }
+
+    const { _id, name, numbers, whatsapps } = profile;
+    const data = { id: _id, name, numbers, whatsapps };
+    return res.status(200).send(data);
   } catch (e) {
+    console.log(e);
     return res.status(500).send();
   }
 };
 
 const updateProfile = async (req, res) => {
   const { id } = req.params;
+  const { username } = req;
 
   try {
-    const profile = await profileService.updateProfile(id, req.body);
+    let profile = await profileService.findProfile(id);
 
-    if (profile) {
-      const { _id: id, name, numbers, whatsapps } = profile;
-      const data = { id, name, numbers, whatsapps };
-      return res.status(200).send(data);
-    } else {
+    if (!profile) {
       return res.status(404).send({ error: 'profile does not exists' });
     }
+
+    const { owner } = profile;
+
+    if (owner !== username) {
+      return res.status(401).send({ error: 'unauthorized' });
+    }
+
+    profile = await profileService.updateProfile(id, req.body);
+
+    const { _id, name, numbers, whatsapps } = profile;
+    const data = { id: _id, name, numbers, whatsapps };
+    return res.status(200).send(data);
   } catch (e) {
     console.log(e);
     return res.status(422).send();
@@ -52,9 +62,20 @@ const updateProfile = async (req, res) => {
 
 const deleteProfile = async (req, res) => {
   const { id } = req.params;
-  console.log(req.username);
+  const { username } = req;
 
   try {
+    const profile = await profileService.findProfile(id);
+
+    if (!profile) {
+      return res.status(200).send();
+    }
+
+    const { owner } = profile;
+    if (owner !== username) {
+      return res.status(401).send({ error: 'unauthorized' });
+    }
+
     await profileService.deleteProfile(id);
     return res.status(200).send();
   } catch (e) {
